refactor(server): layer inference workers through repository factory
The big slice. server/src/lib/inference-client.ts hosted seven loose
functions (listModels, deleteModel, pullModel, importPiperModelToInference,
fetchModelExport, generateAudio, generateAudioStream) that every caller
re-wired into a target object. Replaced by a per-target factory that
keeps the universal-client surface honest:

  server/src/repositories/inference.repository.ts
    inferenceRepository(target) returns a typed client bound to that
    worker. Methods: listModels, deleteModel, fetchExport, pullModel
    (async generator over the upstream SSE), importPiperModel, generate,
    generateStream, transcribe, health. Carries its own typed errors
    (UpstreamError, CacheMissError) instead of plain Error.

  server/src/services/server-models.service.ts
  server/src/services/inference-server.service.ts
  server/src/services/model.service.ts
    Switch every call from `lib/inference-client` to
    `inferenceRepository(target).<method>()`. The health probe in
    inference-server.service collapses to `await repo.health()` and
    drops its inline fetch.

  server/src/lib/inference-router.ts
    Imports InferenceTarget from the repositories barrel now that the
    type lives with the repository.

Routes follow:
  - voice-designer.routes.ts: inferenceRepository(target).generate(...)
  - generate.routes.ts: same; the streaming path also caches the repo
    instance so the cache-miss retry doesn't re-create it.
  - model.routes.ts: inferenceRepository(target).fetchExport(...)
  - transcribe.routes.ts: drops its hand-rolled fetch, calls
    inferenceRepository(target).transcribe(formData, signal); error
    mapping goes through mapServiceError with a 504 override on
    timeouts.

server/src/lib/inference-client.ts removed.
