--- server/src/lib/inference-client.ts
@@ -1,202 +1,0 @@
-import type { CatalogModel } from '@sirene/shared';
-
-export interface InferenceTarget {
-  url: string;
-  authToken?: string;
-}
-
-export interface InferenceRequest {
-  backend: string;
-  text: string;
-  modelPath: string;
-  voicePath?: string;
-  referenceAudio?: string[];
-  referenceAudioData?: string[]; // base64 data URIs, sent on cache miss retry
-  referenceCacheKey?: string;
-  referenceText?: string[];
-  instructText?: string;
-  instructGender?: string;
-  speed?: number;
-  /** Generation-level noise / variation. Only Piper consumes this today (maps to noise_scale). */
-  noiseScale?: number;
-  language?: string;
-}
-
-export class CacheMissError extends Error {
-  public constructor() {
-    super('Reference audio cache miss');
-    this.name = 'CacheMissError';
-  }
-}
-
-export interface StreamingAudioResponse {
-  body: ReadableStream<Uint8Array>;
-  sampleRate: number;
-}
-
-function authHeaders(target: InferenceTarget): Record<string, string> {
-  return target.authToken ? { Authorization: `Bearer ${target.authToken}` } : {};
-}
-
-function buildInferenceBody(request: InferenceRequest) {
-  return {
-    backend: request.backend,
-    text: request.text,
-    model_path: request.modelPath,
-    voice_path: request.voicePath ?? null,
-    reference_audio: request.referenceAudio ?? null,
-    reference_audio_data: request.referenceAudioData ?? null,
-    reference_cache_key: request.referenceCacheKey ?? null,
-    reference_text: request.referenceText ?? null,
-    instruct_text: request.instructText ?? null,
-    instruct_gender: request.instructGender ?? null,
-    speed: request.speed ?? 1.0,
-    noise_scale: request.noiseScale ?? null,
-    language: request.language ?? 'en',
-  };
-}
-
-export async function getModels(target: InferenceTarget): Promise<{ installed: string[]; custom: CatalogModel[] }> {
-  // Throws on transport / non-OK so callers can distinguish a genuinely empty
-  // worker from a transient probe failure (and avoid caching the failure as
-  // "empty" for the TTL window).
-  const response = await fetch(`${target.url}/models`, {
-    headers: authHeaders(target),
-    signal: AbortSignal.timeout(10_000),
-  });
-  if (!response.ok) {
-    throw new Error(`getModels HTTP ${response.status}`);
-  }
-  return response.json() as Promise<{ installed: string[]; custom: CatalogModel[] }>;
-}
-
-export async function deleteModel(target: InferenceTarget, modelId: string): Promise<void> {
-  const response = await fetch(`${target.url}/models/${encodeURIComponent(modelId)}`, {
-    method: 'DELETE',
-    headers: authHeaders(target),
-    signal: AbortSignal.timeout(10_000),
-  });
-  if (!response.ok) {
-    const body = await response.text();
-    throw new Error(`Delete model failed (${response.status}): ${body}`);
-  }
-}
-
-export interface PullModelOptions {
-  backend: string;
-  modelId: string;
-  files: { url: string; path: string }[];
-  totalSize: number;
-  hfToken?: string;
-}
-
-export async function* pullModel(target: InferenceTarget, options: PullModelOptions): AsyncGenerator<Record<string, unknown>> {
-  const response = await fetch(`${target.url}/models/pull`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json', ...authHeaders(target) },
-    body: JSON.stringify({
-      backend: options.backend,
-      model_id: options.modelId,
-      files: options.files,
-      total_size: options.totalSize,
-      hf_token: options.hfToken ?? null,
-    }),
-    signal: AbortSignal.timeout(60 * 60 * 1000), // 1 hour
-  });
-
-  if (!response.ok || !response.body) {
-    const body = await response.text();
-    throw new Error(`Pull model failed (${response.status}): ${body}`);
-  }
-
-  const reader = response.body.getReader();
-  const decoder = new TextDecoder();
-  let buffer = '';
-
-  while (true) {
-    const { done, value } = await reader.read();
-    if (done) {
-      break;
-    }
-    buffer += decoder.decode(value, { stream: true });
-    const lines = buffer.split('\n');
-    buffer = lines.pop() ?? '';
-    for (const line of lines) {
-      if (line.startsWith('data: ')) {
-        try {
-          yield JSON.parse(line.slice(6));
-        } catch {
-          // ignore malformed events
-        }
-      }
-    }
-  }
-}
-
-export async function importPiperModelToInference(target: InferenceTarget, formData: FormData): Promise<{ id: string; message: string }> {
-  const response = await fetch(`${target.url}/models/piper/import`, {
-    method: 'POST',
-    headers: authHeaders(target),
-    body: formData,
-    signal: AbortSignal.timeout(30_000),
-  });
-  if (!response.ok) {
-    const body = (await response.json().catch(() => ({ detail: response.statusText }))) as { detail?: string };
-    const err = Object.assign(new Error(body.detail ?? 'Import failed'), { status: response.status });
-    throw err;
-  }
-  return response.json() as Promise<{ id: string; message: string }>;
-}
-
-export async function fetchModelExport(target: InferenceTarget, modelId: string): Promise<Response> {
-  return fetch(`${target.url}/models/${encodeURIComponent(modelId)}/export`, {
-    headers: authHeaders(target),
-    signal: AbortSignal.timeout(60_000),
-  });
-}
-
-export async function generateAudio(target: InferenceTarget, request: InferenceRequest): Promise<Buffer> {
-  const response = await fetch(`${target.url}/generate`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json', ...authHeaders(target) },
-    body: JSON.stringify(buildInferenceBody(request)),
-  });
-
-  if (response.status === 412) {
-    throw new CacheMissError();
-  }
-
-  if (!response.ok) {
-    const body = await response.text();
-    throw new Error(`Inference failed (${response.status}): ${body}`);
-  }
-
-  const arrayBuffer = await response.arrayBuffer();
-  return Buffer.from(arrayBuffer);
-}
-
-export async function generateAudioStream(target: InferenceTarget, request: InferenceRequest): Promise<StreamingAudioResponse> {
-  const response = await fetch(`${target.url}/generate/stream`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json', ...authHeaders(target) },
-    body: JSON.stringify(buildInferenceBody(request)),
-  });
-
-  if (response.status === 412) {
-    throw new CacheMissError();
-  }
-
-  if (!response.ok) {
-    const body = await response.text();
-    throw new Error(`Inference streaming failed (${response.status}): ${body}`);
-  }
-
-  if (!response.body) {
-    throw new Error('No response body for streaming');
-  }
-
-  return {
-    body: response.body as ReadableStream<Uint8Array>,
-    sampleRate: Number(response.headers.get('X-Sample-Rate') ?? '24000'),
-  };
-}
