+++ server/src/repositories/index.ts
@@ -1,5 +1,6 @@
 export { elevenlabsRepository } from './elevenlabs.repository';
 export { generationRepository } from './generation.repository';
+export { CacheMissError, type InferenceRequest, type InferenceTarget, inferenceRepository, type PullModelOptions, type StreamingAudioResponse } from './inference.repository';
 export { inferenceServerRepository } from './inference-server.repository';
 export { OPENAI_VOICES, openAITtsRepository } from './openai-tts.repository';
 export { sessionRepository } from './session.repository';
